Fix internal repo clone to dir with absolute path
`env.WORKSPACE` is not guaranteed to be the same across slaves within
the same build. If the directory, in which to clone the repository,
includes the workspace this will cause the clone to fail if the value of
`env.WORKSPACE` for the parent slave does not match the internal slave
where the clone is done. This change strips the workspace from the
directory path used by the internal slave, if included in
the directory path supplied to the function.

JIRA: RE-2338