Fix build summary naming issue
When the Build-Docker-Images-For-Master is ran, it uploads using a suffix
'build-summary'. The Build-Summary's job's default name is 'buildsummary'.
This results in a naming mismatch with Build-Summary causing it to fail. When
manually specifying build-summary, it completes without issue.  This change
updates the default value to match what is actually being uploaded to the docker
registry.