Jenkins upkeep
Venv builds fail due to insecure cert leaving
us not able to fix it via jenkins(chicken-egg).
I'm moving the REPO_BASE to http to see if it
works around the issue.

While in the process of testing, I found the
public cloud image for some of the testing has
changed its name.